+++ app/src/main/AndroidManifest.xml
@@ -33,6 +33,8 @@
         </activity>
         <activity android:name=".GameActivity">
         </activity>
+        <activity android:name=".GameActivity_A">
+        </activity>
     </application>
 
 </manifest>
