+++ app/src/main/java/com/example/imsu/GameActivity_B.java
@@ -1,0 +1,4 @@
+package com.example.imsu;
+
+public class GameActivity_B {
+}
