+++ app/src/main/java/com/example/imsu/LevelManager.java
@@ -1,0 +1,12 @@
+package com.example.imsu;
+
+public class LevelManager {
+    public void getLevel(Player player) {
+
+    }
+
+    public void saveProgress() {
+
+    }
+
+}
