+++ app/src/main/res/layout/activity_play.xml
@@ -12,32 +12,32 @@
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
-        android:layout_marginTop="80dp"
         android:layout_marginEnd="16dp"
         android:background="@drawable/drawable_button_rounded"
         android:text="@string/btn_Continue"
-        android:textColor="@color/colorTextLight"
-        android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView" />
-
-    <Button
-        android:id="@+id/btn_NewGame"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginEnd="16dp"
-        android:background="@drawable/drawable_button_rounded"
-        android:text="@string/btn_NewGame"
         android:textColor="@color/colorTextLight"
         android:textSize="24sp"
         app:layout_constraintBottom_toTopOf="@+id/btn_SelectLevel"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btn_Continue" />
+        app:layout_constraintTop_toBottomOf="@+id/btn_NewGame" />
+
+    <Button
+        android:id="@+id/btn_NewGame"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="80dp"
+        android:layout_marginEnd="16dp"
+        android:background="@drawable/drawable_button_rounded"
+        android:text="@string/btn_NewGame"
+        android:textColor="@color/colorTextLight"
+        android:textSize="24sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textView" />
 
     <Button
         android:id="@+id/btn_SelectLevel"
